Extract common int prim code
